기존 repository 코드를 common repository 로 재배치.
다른 repository 들이 common 모듈을 불러와 사용하기 위한 전략.
