+++ repository/common/src/main/java/com/season/winter/common/dataSet/liquorInfo/ScotchLiquorInfo.kt
@@ -1,4 +1,4 @@
-package com.season.winter.liquor.dummy.dataSet.liquorInfo
+package com.season.winter.common.dataSet.liquorInfo
 
 import com.season.winter.liquor.liquorInfo.LiquorExpressionCode
 import com.season.winter.liquor.liquorInfo.LiquorInfo
@@ -7,11 +7,11 @@ import com.season.winter.liquor.liquorInfo.WhiskyType
 import com.season.winter.liquor.brand.BrandInfo
 import com.season.winter.liquor.content.FeatureData
 import com.season.winter.liquor.content.FeatureType
-import com.season.winter.liquor.dummy.dummy.getScotchLiquorIdentify
+import com.season.winter.common.dummy.getScotchLiquorIdentify
 import com.season.winter.liquor.liquorInfo.LiquorStatus
 
 val theMacllan12ySherryOak = LiquorInfo(
-    identify = getScotchLiquorIdentify(0),
+    identify = com.season.winter.common.dummy.getScotchLiquorIdentify(0),
     nameKo = "맥켈란 12년 셰리 오크",
     nameEn = "The Macllan 12y Sherry Oak",
     wishCount = 0,
@@ -70,7 +70,7 @@ val theMacllan12ySherryOak = LiquorInfo(
 )
 
 val balvenieDoublewood12Yo = LiquorInfo(
-    identify = getScotchLiquorIdentify(1),
+    identify = com.season.winter.common.dummy.getScotchLiquorIdentify(1),
     nameKo = "발베니 더블우드 12년",
     nameEn = "Balvenie Doublewood 12yo",
     wishCount = 0,
@@ -122,7 +122,7 @@ val balvenieDoublewood12Yo = LiquorInfo(
 )
 
 val glenmorangieLasanta12yo = LiquorInfo(
-    identify = getScotchLiquorIdentify(2),
+    identify = com.season.winter.common.dummy.getScotchLiquorIdentify(2),
     nameKo = "글렌모렌지 라산타 12년",
     nameEn = "Glenmorangie Lasanta 12yo",
     wishCount = 0,
