+++ repository/common/src/main/java/com/season/winter/common/dummy/LiquorShopDummyData.kt
@@ -1,0 +1,2 @@
+package com.season.winter.common.dummy
+
