--- repository/liquor/src/main/java/com/season/winter/liquor/dummy/dummy/LiquorShopDummyData.kt
@@ -1,2 +1,0 @@
-package com.season.winter.liquor.dummy.dummy
-
